Cleaner Code and TODO Removal
In the class files for the ORM I have removed the TODO's as it is implemented
properly. It can always be made better if needed but it meets needs right now

I have also renamed the main file OpenODSDataSetup which contains a class of
same name. This is the orchestration class really and will be called by our
main.py somewhere.