Models, Controller and Addresses have been added
I have moved the models into a model folder so we can use them for reads later.
I have also created a blank database controller to start working on, this does
nothing at the minute.
I have added the addresses model, this is going to break the reads as I have
renamed some of the table columns but it won't matter when we refactor the
front end to use these models too.