ODSFile Managment has been seperated out into it's own class
A new class now handles file retrieval and extraction. Unit tests have been
created that now need to be made to pass. They are not exhaustive at this stage
but it gives a better understanding of how we need to structure the code and
the funcitons we are building.

Also as part of this commit I created a settings model, however I am not sure
of the purpose of this for the project going forward.