Datafile is automatically downloaded and extracted fixes issue#42
This happens without knowledge of the contents of the zipfile and the
internal file name. This has been fixed so we can always have the latest
data available and build a cron job to automate the process.

Also some of the more recent commits were not PEP8 compliant so they have been
fixed but that is a style thing. No difference in functionality.